Time: 2 ms (97.86%), Space: 48.5 MB (59.91%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.java
@@ -19,15 +19,15 @@ class Solution {
         Queue<TreeNode> queue=new LinkedList<>();
         queue.add(root);
         while(!queue.isEmpty()){
-            Double avg=0.00000;
+            long avg=0;
             int size=queue.size();
             for(int i=0;i<size;i++){
                 TreeNode node=queue.poll();
-                avg+=(double)node.val;
+                avg+=node.val;
                 if(node.left!=null) queue.add(node.left);
                 if(node.right!=null) queue.add(node.right);
             }
-            result.add(avg/size);
+            result.add((double)avg/size);
         }
         return result;
     }
